Recreate test project
