+++ BistroQ.Tests.MSTest/Test1.cs
@@ -1,0 +1,35 @@
+﻿namespace BistroQ.Tests.MsTest
+{
+    [TestClass]
+    public sealed class Test1
+    {
+        [AssemblyInitialize]
+        public static void AssemblyInit(TestContext context)
+        {
+            // This method is called once for the test assembly, before any tests are run.
+        }
+
+        [AssemblyCleanup]
+        public static void AssemblyCleanup()
+        {
+            // This method is called once for the test assembly, after all tests are run.
+        }
+
+        [TestInitialize]
+        public void TestInit()
+        {
+            // This method is called before each test method.
+        }
+
+        [TestCleanup]
+        public void TestCleanup()
+        {
+            // This method is called after each test method.
+        }
+
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
