--- BistroQ.Tests.MSTest/Usings.cs
@@ -1,2 +1,0 @@
-﻿global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using Microsoft.VisualStudio.TestTools.UnitTesting.AppContainer;
