+++ BistroQ.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35327.3
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BistroQ.Tests.MSTest", "BistroQ.Tests.MSTest\BistroQ.Tests.MSTest.csproj", "{1DB31CE5-6A73-4D04-B883-1466323E677F}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BistroQ.Domain", "BistroQ.Domain\BistroQ.Domain.csproj", "{40CF8F26-9753-44EE-A809-8E978D972BEA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BistroQ.Presentation", "BistroQ.Presentation\BistroQ.Presentation.csproj", "{E825B041-B45A-4E3E-8CC4-C71BA277762F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BistroQ.Service", "BistroQ.Service\BistroQ.Service.csproj", "{002A09C7-5E19-44F0-9146-FAAF432062D3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BistroQ.Tests.MsTest", "BistroQ.Tests.MsTest\BistroQ.Tests.MsTest.csproj", "{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,22 +23,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|Any CPU.Build.0 = Debug|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|arm64.ActiveCfg = Debug|arm64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|arm64.Build.0 = Debug|arm64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|x64.ActiveCfg = Debug|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|x64.Build.0 = Debug|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|x86.ActiveCfg = Debug|x86
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Debug|x86.Build.0 = Debug|x86
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|Any CPU.ActiveCfg = Release|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|Any CPU.Build.0 = Release|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|arm64.ActiveCfg = Release|arm64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|arm64.Build.0 = Release|arm64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|x64.ActiveCfg = Release|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|x64.Build.0 = Release|x64
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|x86.ActiveCfg = Release|x86
-		{1DB31CE5-6A73-4D04-B883-1466323E677F}.Release|x86.Build.0 = Release|x86
 		{40CF8F26-9753-44EE-A809-8E978D972BEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{40CF8F26-9753-44EE-A809-8E978D972BEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{40CF8F26-9753-44EE-A809-8E978D972BEA}.Debug|arm64.ActiveCfg = Debug|arm64
@@ -95,6 +79,22 @@ Global
 		{002A09C7-5E19-44F0-9146-FAAF432062D3}.Release|x64.Build.0 = Release|Any CPU
 		{002A09C7-5E19-44F0-9146-FAAF432062D3}.Release|x86.ActiveCfg = Release|Any CPU
 		{002A09C7-5E19-44F0-9146-FAAF432062D3}.Release|x86.Build.0 = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|arm64.ActiveCfg = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|arm64.Build.0 = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|x64.Build.0 = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Debug|x86.Build.0 = Debug|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|arm64.ActiveCfg = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|arm64.Build.0 = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|x64.ActiveCfg = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|x64.Build.0 = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|x86.ActiveCfg = Release|Any CPU
+		{27EEC89C-5772-4FAB-B5E3-0706F6993E4E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
